+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Recalculation of function when result is null in cache of AppBio Quantstudio
 ### Changed
+- Replaced datetime.timezone with ZoneInfo in TimestampParser
 ### Deprecated
 ### Removed
 ### Security